Vastly improved the camera and sprite rendering solution.
Now the camera is centered to the window, sprites are in 'world coordinates' although the units are still pixels. Sprites now draw from bottom-left to top-right by increasing position, like many other rendering engines. Sprites now have support for pivot points, which is where they are rendered around. Pivot of (0, 0) means that the position of the sprite is the bottom-left of the image, pivot (0.5, 0.5) means that the sprite position is the center of the image and pivot (1, 1) is the top-right of the image.